+++ examples/basic/shadow1.py
@@ -2,9 +2,9 @@
 from vedo import dataurl, Mesh, Sphere, show
 
 spider = Mesh(dataurl+"spider.ply")
+spider.normalize().rotate_z(-90)
 spider.texture(dataurl+'textures/leather.jpg')
-
-spider.normalize().rotate_z(-90).add_shadow('x', -3, alpha=0.5)
+spider.add_shadow('x', -3)
 
 sphere = Sphere(r=0.3).pos(0.4,0,0.6).add_shadow('x', -3)