+++ examples/simulations/pendulum_3d.py
@@ -1,6 +1,5 @@
-"""Double pendulum in 3D
-Press ESC to quit"""
-# Original idea and solution using sympy from Mr. P Solver at:
+"""Double pendulum in 3D (press ESC to quit)"""
+# Original idea and solution using sympy from:
 # https://www.youtube.com/watch?v=MtG9cueB548
 from vedo import *
 
@@ -8,8 +7,14 @@ from vedo import *
 x1, y1, z1,  x2, y2, z2 = np.load(download(dataurl+'3Dpen.npy'))
 p1, p2 = np.c_[x1,y1,z1], np.c_[x2,y2,z2]
 
-ball1 = Sphere(c="green5", r=0.1, pos=p1[0]).add_trail(n=10).add_shadow('z', -3)
-ball2 = Sphere(c="blue5",  r=0.1, pos=p2[0]).add_trail(n=10).add_shadow('z', -3)
+ball1 = Sphere(c="green5", r=0.1, pos=p1[0])
+ball2 = Sphere(c="blue5",  r=0.1, pos=p2[0])
+
+ball1.add_shadow('z', -3)
+ball2.add_shadow('z', -3)
+
+ball1.add_trail(n=20)
+ball2.add_trail(n=20)
 ball1.trail.add_shadow('z', -3) # make trails project a shadow too
 ball2.trail.add_shadow('z', -3)
 
@@ -18,11 +23,12 @@ rod2 = Line(ball1, ball2, lw=4)
 
 axes = Axes(xrange=(-3,3), yrange=(-3,3), zrange=(-3,3))
 
+# show the solution
 plt = Plotter(interactive=False)
 plt.show(axes, __doc__, viewup='z')
 
 i = 0
-for b1, b2 in progressbar(zip(p1,p2)):
+for b1, b2 in zip(p1,p2):
     ball1.pos(b1)
     ball2.pos(b2)
     rod1.stretch([0,0,0], b1)
@@ -32,6 +38,5 @@ for b1, b2 in progressbar(zip(p1,p2)):
     i+=1
     if i > 50:
         break
-print("\ndone.")
 
 plt.interactive().close()