+++ docs/changes.md
@@ -27,7 +27,7 @@
 - Add `application.Browser().play()` to autoplay a slider
 - Add `pad()` to padding a Volume with zeros voxels (useful to `dilate()`)
 
-
+- Add `ProgressBarWindow()` to show a progress bar in the rendering window
 
 -------------------------
 ## New/Revised Examples
@@ -38,6 +38,7 @@ examples/advanced/timer_callback3.py
 examples/advanced/warp6.py
 examples/pyplot/histo_1d_e.py
 examples/other/tensor_grid2.py
+examples/simulations/lorentz.py
 ```
 
 ### Broken Examples